requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
isort==5.12.0
pyTelegramBotAPI==4.10.0
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
python-dotenv==0.21.1